allow to buy shields even if you already have one

+++ include/item.h
@@ -78,5 +78,5 @@ namespace mod::item
 	/**
 	 * Contains the values for the flags to be set to skip the animations of first getting specific items
 	 */
-	extern u16 itemGetAnimationFlags[10];
+	extern u8 itemGetAnimationFlags[10];
 }
